+++ instrumentation/netty/netty-4.1/library/src/main/java/io/opentelemetry/instrumentation/netty/v4_1/internal/ServerContext.java
@@ -7,7 +7,7 @@ package io.opentelemetry.instrumentation.netty.v4_1.internal;
 
 import com.google.auto.value.AutoValue;
 import io.opentelemetry.context.Context;
-import io.opentelemetry.instrumentation.netty.v4.common.HttpRequestAndChannel;
+import io.opentelemetry.instrumentation.netty.common.v4_0.HttpRequestAndChannel;
 
 /**
  * A tuple of an {@link Context} and a {@link HttpRequestAndChannel}.
