+++ instrumentation/netty/netty-common-4.0/library/src/main/java/io/opentelemetry/instrumentation/netty/common/v4_0/internal/client/NettyErrorOnlyConnectionInstrumenter.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.netty.v4.common.internal.client;
+package io.opentelemetry.instrumentation.netty.common.v4_0.internal.client;
 
 import io.netty.channel.Channel;
 import io.opentelemetry.context.Context;
