+++ instrumentation/netty/netty-common-4.0/library/src/main/java/io/opentelemetry/instrumentation/netty/common/v4_0/internal/server/NettyHttpServerAttributesGetter.java
@@ -3,45 +3,25 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.netty.v4.common.internal.client;
-
-import static io.opentelemetry.instrumentation.netty.v4.common.internal.HttpSchemeUtil.getScheme;
+package io.opentelemetry.instrumentation.netty.common.v4_0.internal.server;
 
 import io.netty.handler.codec.http.HttpResponse;
 import io.netty.handler.codec.http.HttpVersion;
-import io.opentelemetry.instrumentation.api.semconv.http.HttpClientAttributesGetter;
-import io.opentelemetry.instrumentation.netty.v4.common.HttpRequestAndChannel;
-import io.opentelemetry.instrumentation.netty.v4.common.internal.ChannelUtil;
+import io.opentelemetry.instrumentation.api.semconv.http.HttpServerAttributesGetter;
+import io.opentelemetry.instrumentation.netty.common.v4_0.HttpRequestAndChannel;
+import io.opentelemetry.instrumentation.netty.common.v4_0.internal.ChannelUtil;
+import io.opentelemetry.instrumentation.netty.common.v4_0.internal.HttpSchemeUtil;
 import java.net.InetSocketAddress;
 import java.net.SocketAddress;
-import java.net.URI;
-import java.net.URISyntaxException;
 import java.util.List;
 import javax.annotation.Nullable;
 
-final class NettyHttpClientAttributesGetter
-    implements HttpClientAttributesGetter<HttpRequestAndChannel, HttpResponse> {
-
-  @Override
-  @Nullable
-  public String getUrlFull(HttpRequestAndChannel requestAndChannel) {
-    try {
-      String hostHeader = getHost(requestAndChannel);
-      String target = requestAndChannel.request().getUri();
-      URI uri = new URI(target);
-      if ((uri.getHost() == null || uri.getHost().equals("")) && hostHeader != null) {
-        return getScheme(requestAndChannel) + "://" + hostHeader + target;
-      }
-      return uri.toString();
-    } catch (URISyntaxException e) {
-      return null;
-    }
-  }
-
-  private String getHost(HttpRequestAndChannel requestAndChannel) {
-    List<String> values = getHttpRequestHeader(requestAndChannel, "host");
-    return values.isEmpty() ? null : values.get(0);
-  }
+/**
+ * This class is internal and is hence not for public use. Its APIs are unstable and can change at
+ * any time.
+ */
+public final class NettyHttpServerAttributesGetter
+    implements HttpServerAttributesGetter<HttpRequestAndChannel, HttpResponse> {
 
   @Override
   public String getHttpRequestMethod(HttpRequestAndChannel requestAndChannel) {
@@ -66,8 +46,27 @@ final class NettyHttpClientAttributesGetter
   }
 
   @Override
+  public String getUrlScheme(HttpRequestAndChannel requestAndChannel) {
+    return HttpSchemeUtil.getScheme(requestAndChannel);
+  }
+
+  @Override
+  public String getUrlPath(HttpRequestAndChannel requestAndChannel) {
+    String fullPath = requestAndChannel.request().getUri();
+    int separatorPos = fullPath.indexOf('?');
+    return separatorPos == -1 ? fullPath : fullPath.substring(0, separatorPos);
+  }
+
+  @Override
+  public String getUrlQuery(HttpRequestAndChannel requestAndChannel) {
+    String fullPath = requestAndChannel.request().getUri();
+    int separatorPos = fullPath.indexOf('?');
+    return separatorPos == -1 ? null : fullPath.substring(separatorPos + 1);
+  }
+
+  @Override
   public String getNetworkTransport(
-      HttpRequestAndChannel requestAndChannel, @Nullable HttpResponse response) {
+      HttpRequestAndChannel requestAndChannel, HttpResponse response) {
     return ChannelUtil.getNetworkTransport(requestAndChannel.channel());
   }
 
@@ -87,23 +86,22 @@ final class NettyHttpClientAttributesGetter
     return version.majorVersion() + "." + version.minorVersion();
   }
 
-  @Nullable
-  @Override
-  public String getServerAddress(HttpRequestAndChannel requestAndChannel) {
-    return null;
-  }
-
-  @Nullable
-  @Override
-  public Integer getServerPort(HttpRequestAndChannel requestAndChannel) {
-    return null;
-  }
-
   @Override
   @Nullable
   public InetSocketAddress getNetworkPeerInetSocketAddress(
       HttpRequestAndChannel requestAndChannel, @Nullable HttpResponse response) {
     SocketAddress address = requestAndChannel.remoteAddress();
+    if (address instanceof InetSocketAddress) {
+      return (InetSocketAddress) address;
+    }
+    return null;
+  }
+
+  @Nullable
+  @Override
+  public InetSocketAddress getNetworkLocalInetSocketAddress(
+      HttpRequestAndChannel requestAndChannel, @Nullable HttpResponse response) {
+    SocketAddress address = requestAndChannel.channel().localAddress();
     if (address instanceof InetSocketAddress) {
       return (InetSocketAddress) address;
     }
